rename listener_resource_type.go to helpers_test.go
